lvconvert: add lvconvert_to_pool_or_swap_metadata
Give this syntax:

lvconvert --thinpool LV1 --poolmetadata LV2

To two different commands:

1. Convert LV1 to a pool when LV1 is not yet a pool.
   In this case the metadata LV is optional.

2. Swap the metadata LV with LV2 when LV1 is already a pool.
   In this case the metadata LV is required.

This makes the syntax ambiguous; we do not know by looking
at it what the command will do.  I believe it is very
unwise and unnecessary to make this syntax ambiguous.

Given that 1 is a routine command, and 2 is a dangerous
command that should not be used in general, and has been widely
frowned upon, conflating these two commands into one syntax
creates dangerous confusion to users.  There is really no need
for this to swap metadata LVs.  Both of the two operations
conflated have unique and umambiguous primary commands:

1. Primary form of creating a thin pool:
lvconvert --type thin-pool --poolmetadata LV2 LV1

2. Primary form of swapping metadata LV:
lvconvert --swapmetadata --poolmetadata LV2 LV1

So we're talking here about which of these primary commands
should be given the alternate syntax
(lvconvert --poolmetadata --thinpool), or whether the
alternate syntax should be given to both of them.

(The same case applies to lvconvert --cachepool which is also
included here.)