lvmdbusd: Replay flight recorder json data
This functionality allows the ability to take a directory and
place json files in it with the name N.json where N starts at 0
and increments.  To get the service to consume each file you can
invoke the ExternalEvent dbus method on the manager interface.

To use set env. variable LVMDBUSD_REPLAY to the directory that
contains the lvm json files.  This can be used for instances where
we have the flight recorder data and we want to replay the sequence
for debugging.

Signed-off-by: Tony Asleson <tasleson@redhat.com>